pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "bzsplatting"
version = "0.1.0"
description = "Pedagogical reproduction of Bézier Splatting for differentiable vector graphics"
authors = [{name = "OpenAI Assistant"}]
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
    "torch>=2.0",
    "Pillow>=10.0",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "-ra"
